pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "google_stopwatch"
version = "0.0.1"
authors = [{ name = "Rafay Khan" }]
maintainers = [{ name = "Rafay Khan" }]
description = "A stopwatch program that automatically logs sessions to Google Calendar."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "pyyaml",
]

[project.optional-dependencies]
development = ["black", "ruff", "pytype"]

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP", # pyupgrade
    "D",  # pydocstyle
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
